move bellatrix block generator
